+++ data-index/data-index-service/data-index-service-common/src/test/java/org/kie/kogito/index/messaging/AbstractDomainMessagingConsumerKafkaIT.java
@@ -27,7 +27,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.kie.kogito.index.DataIndexStorageService;
 import org.kie.kogito.persistence.protobuf.ProtobufService;
-import org.kie.kogito.test.kafka.KafkaTestClient;
+import org.kie.kogito.test.quarkus.kafka.KafkaTestClient;
 import org.kie.kogito.testcontainers.quarkus.KafkaQuarkusTestResource;
 
 import io.restassured.http.ContentType;
